fix #9: Read timestamps from idx files
To my curernt understanding, the time logs are saved in the '_idx.bin'
files as time delta in nanoseconds (Not as absolute time - Maybe as
absolute in the reference of start time of the EVM board).

Thus, the saved timestamps use the time at the moment of repacking the
recordings and ajust them with the time logs from the EVM board.